automll_package/requirements.txt
seaborn
pandas
numpy
ipywidgets
scikit-learn
pandas_profiling
streamlit_pandas_profiling
streamlit_extras
